refactor: 전체 리팩토링 및 패키지 이동
 원시값 포장 및 불필요 코드 제거
