+++ src/main/java/chess/domain/piece/Bishop.java
@@ -6,35 +6,40 @@ import java.util.stream.Collectors;
 import chess.domain.position.Position;
 
 public class Bishop extends Piece {
-    public Bishop(Position position, Team team) {
-        super(position, Name.BISHOP, team);
-    }
+	public Bishop(Position position, Team team) {
+		super(position, Name.BISHOP, team);
+	}
 
-    @Override
-    public boolean canNotMoveTo(Piece that) {
-        return isSameTeam(that.team) || !createMovableArea().contains(that.position);
-    }
+	@Override
+	public boolean canNotMoveTo(Piece that) {
+		return isSameTeam(that.team) || !createMovableArea().contains(that.position);
+	}
 
-    @Override
-    protected List<Position> createMovableArea() {
-        return Position.getPositions()
-                .stream()
-                .filter(position -> !position.equals(this.position))
-                .filter(this::isDiagonal)
-                .collect(Collectors.toList());
-    }
+	@Override
+	protected List<Position> createMovableArea() {
+		return Position.getPositions()
+			.stream()
+			.filter(position -> !position.equals(this.position))
+			.filter(this::isDiagonal)
+			.collect(Collectors.toList());
+	}
 
-    @Override
-    public boolean isObstacle() {
-        return true;
-    }
+	private boolean isDiagonal(Position position) {
+		return Math.abs(position.getColumnGap(this.position)) == Math.abs(position.getRowGap(this.position));
+	}
 
-    @Override
-    public boolean hasToAlive() {
-        return false;
-    }
+	@Override
+	public boolean isObstacle() {
+		return true;
+	}
 
-    private boolean isDiagonal(Position position) {
-        return Math.abs(position.getColumnGap(this.position)) == Math.abs(position.getRowGap(this.position));
-    }
+	@Override
+	public boolean hasToAlive() {
+		return false;
+	}
+
+	@Override
+	public boolean isPenaltyApplier() {
+		return false;
+	}
 }
