+++ src/main/java/chess/domain/piece/King.java
@@ -6,35 +6,43 @@ import java.util.stream.Collectors;
 import chess.domain.position.Position;
 
 public class King extends Piece {
-    public King(Position position, Team team) {
-        super(position, Name.KING, team);
-    }
+	private static final int AROUND_BOUND = 1;
 
-    @Override
-    public boolean canNotMoveTo(Piece that) {
-        return isSameTeam(that.team) || !createMovableArea().contains(that.position);
-    }
+	public King(Position position, Team team) {
+		super(position, Name.KING, team);
+	}
 
-    @Override
-    protected List<Position> createMovableArea() {
-        return Position.getPositions()
-                .stream()
-                .filter(position -> !position.equals(this.position))
-                .filter(this::isAround)
-                .collect(Collectors.toList());
-    }
+	@Override
+	public boolean canNotMoveTo(Piece that) {
+		return isSameTeam(that.team) || !createMovableArea().contains(that.position);
+	}
 
-    private boolean isAround(Position position) {
-        return Math.abs(position.getColumnGap(this.position)) <= 1 && Math.abs(position.getRowGap(this.position)) <= 1;
-    }
+	@Override
+	protected List<Position> createMovableArea() {
+		return Position.getPositions()
+			.stream()
+			.filter(position -> !position.equals(this.position))
+			.filter(this::isAround)
+			.collect(Collectors.toList());
+	}
 
-    @Override
-    public boolean isObstacle() {
-        return true;
-    }
+	private boolean isAround(Position position) {
+		return Math.abs(position.getColumnGap(this.position)) <= AROUND_BOUND &&
+			Math.abs(position.getRowGap(this.position)) <= AROUND_BOUND;
+	}
 
-    @Override
-    public boolean hasToAlive() {
-        return true;
-    }
+	@Override
+	public boolean isObstacle() {
+		return true;
+	}
+
+	@Override
+	public boolean hasToAlive() {
+		return true;
+	}
+
+	@Override
+	public boolean isPenaltyApplier() {
+		return false;
+	}
 }
