+++ src/main/java/chess/domain/piece/Piece.java
@@ -5,41 +5,43 @@ import java.util.List;
 import chess.domain.position.Position;
 
 public abstract class Piece {
-    protected Position position;
-    protected final Name name;
-    protected final Team team;
+	protected Position position;
+	protected final Name name;
+	protected final Team team;
 
-    public Piece(Position position, Name name, Team team) {
-        this.position = position;
-        this.name = name;
-        this.team = team;
-    }
+	public Piece(Position position, Name name, Team team) {
+		this.position = position;
+		this.name = name;
+		this.team = team;
+	}
 
-    public void moveTo(Position destination) {
-        position = destination;
-    }
+	public void moveTo(Position destination) {
+		position = destination;
+	}
 
-    public abstract boolean canNotMoveTo(Piece that);
+	public abstract boolean canNotMoveTo(Piece that);
 
-    protected abstract List<Position> createMovableArea();
+	protected abstract List<Position> createMovableArea();
 
-    public abstract boolean isObstacle();
+	public abstract boolean isObstacle();
 
-    public abstract boolean hasToAlive();
+	public abstract boolean hasToAlive();
 
-    public boolean isSameTeam(Team team) {
-        return this.team == team;
-    }
+	public abstract boolean isPenaltyApplier();
 
-    public boolean isNotSameTeam(Team team) {
-        return this.team != team;
-    }
+	public boolean isSameTeam(Team team) {
+		return this.team == team;
+	}
 
-    public Position getPosition() {
-        return position;
-    }
+	public boolean isNotSameTeam(Team team) {
+		return this.team != team;
+	}
 
-    public String getName() {
-        return name.getName();
-    }
+	public Position getPosition() {
+		return position;
+	}
+
+	public String getName() {
+		return name.getName();
+	}
 }
