+++ src/main/java/chess/domain/position/Distance.java
@@ -1,23 +1,35 @@
 package chess.domain.position;
 
 public class Distance {
-    private final int distance;
+	private final int distance;
 
-    private Distance(int distance) {
-        this.distance = distance;
-    }
+	private Distance(int distance) {
+		this.distance = distance;
+	}
 
-    public static Distance of(Position start, Position end) {
-        int columnDistance = Math.abs(start.getColumnGap(end));
-        int rowDistance = Math.abs(start.getRowGap(end));
+	public static Distance of(Position start, Position end) {
+		int columnDistance = Math.abs(start.getColumnGap(end));
+		int rowDistance = Math.abs(start.getRowGap(end));
 
-        if (columnDistance == rowDistance || columnDistance == 0 || rowDistance == 0) {
-            return new Distance(Integer.max(columnDistance, rowDistance));
-        }
-        return new Distance(0);
-    }
+		if (isDiagonal(columnDistance, rowDistance) || isVertical(columnDistance) || isHorizontal(rowDistance)) {
+			return new Distance(Integer.max(columnDistance, rowDistance));
+		}
+		return new Distance(0);
+	}
 
-    public int getDistance() {
-        return distance;
-    }
+	private static boolean isDiagonal(int columnDistance, int rowDistance) {
+		return columnDistance == rowDistance;
+	}
+
+	private static boolean isVertical(int columnDistance) {
+		return columnDistance == 0;
+	}
+
+	private static boolean isHorizontal(int rowDistance) {
+		return rowDistance == 0;
+	}
+
+	public int getDistance() {
+		return distance;
+	}
 }
