+++ src/main/java/chess/domain/position/Path.java
@@ -5,25 +5,25 @@ import java.util.Collections;
 import java.util.List;
 
 public class Path {
-    private final List<Position> path;
+	private final List<Position> path;
 
-    private Path(List<Position> path) {
-        this.path = path;
-    }
+	private Path(List<Position> path) {
+		this.path = path;
+	}
 
-    public static Path of(Position start, Position end) {
-        List<Position> path = new ArrayList<>();
+	public static Path of(Position start, Position end) {
+		List<Position> path = new ArrayList<>();
 
-        Position current = start;
-        for (int i = 1; i < Distance.of(start, end).getDistance(); i++) {
-            current = Direction.of(current, end).move(current);
-            path.add(current);
-        }
+		Position current = start;
+		for (int i = 1; i < Distance.of(start, end).getDistance(); i++) {
+			current = Direction.of(current, end).move(current);
+			path.add(current);
+		}
 
-        return new Path(path);
-    }
+		return new Path(path);
+	}
 
-    public List<Position> toList() {
-        return Collections.unmodifiableList(path);
-    }
+	public List<Position> toList() {
+		return Collections.unmodifiableList(path);
+	}
 }
