+++ src/main/java/chess/view/InputView.java
@@ -3,13 +3,13 @@ package chess.view;
 import java.util.Scanner;
 
 public class InputView {
-    private static final Scanner scanner = new Scanner(System.in);
+	private static final Scanner scanner = new Scanner(System.in);
 
-    public static String inputMoveOrStatus() {
-        return scanner.nextLine();
-    }
+	public static String inputMoveOrStatus() {
+		return scanner.nextLine();
+	}
 
-    public static String inputStartOrEnd() {
-        return scanner.nextLine();
-    }
+	public static String inputStartOrEnd() {
+		return scanner.nextLine();
+	}
 }
