+++ src/test/java/chess/domain/piece/KingTest.java
@@ -1,34 +1,34 @@
 package chess.domain.piece;
 
+import static chess.domain.position.Fixtures.*;
+import static org.assertj.core.api.Assertions.*;
+
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import static chess.domain.position.Fixtures.*;
-import static org.assertj.core.api.Assertions.assertThat;
-
 class KingTest {
-    King king;
+	King king;
 
-    @BeforeEach
-    void setUp() {
-        king = new King(C3, Team.BLACK);
-    }
+	@BeforeEach
+	void setUp() {
+		king = new King(C3, Team.BLACK);
+	}
 
-    @Test
-    void moveTo() {
-        king.moveTo(C4);
-        assertThat(king.getPosition()).isEqualTo(C4);
-    }
+	@Test
+	void moveTo() {
+		king.moveTo(C4);
+		assertThat(king.getPosition()).isEqualTo(C4);
+	}
 
-    @Test
-    void canNotMoveTo_Return_True() {
-        Piece target = new Empty(D5);
-        assertThat(king.canNotMoveTo(target)).isTrue();
-    }
+	@Test
+	void canNotMoveTo_Return_True() {
+		Piece target = new Empty(D5);
+		assertThat(king.canNotMoveTo(target)).isTrue();
+	}
 
-    @Test
-    void canNotMoveTo_Return_False() {
-        Piece target = new Empty(C4);
-        assertThat(king.canNotMoveTo(target)).isFalse();
-    }
+	@Test
+	void canNotMoveTo_Return_False() {
+		Piece target = new Empty(C4);
+		assertThat(king.canNotMoveTo(target)).isFalse();
+	}
 }
