+++ src/test/java/chess/domain/position/Fixtures.java
@@ -1,40 +1,40 @@
 package chess.domain.position;
 
 public class Fixtures {
-    public static final Position A1 = Position.of("a1");
-    public static final Position A2 = Position.of("a2");
-    public static final Position A3 = Position.of("a3");
-    public static final Position A4 = Position.of("a4");
-    public static final Position A5 = Position.of("a5");
-    public static final Position B1 = Position.of("b1");
-    public static final Position B2 = Position.of("b2");
-    public static final Position B3 = Position.of("b3");
-    public static final Position B4 = Position.of("b4");
-    public static final Position B5 = Position.of("b5");
-    public static final Position B6 = Position.of("b6");
-    public static final Position C1 = Position.of("c1");
-    public static final Position C2 = Position.of("c2");
-    public static final Position C3 = Position.of("c3");
-    public static final Position C4 = Position.of("c4");
-    public static final Position C5 = Position.of("c5");
-    public static final Position C6 = Position.of("c6");
-    public static final Position C7 = Position.of("c7");
-    public static final Position C8 = Position.of("c8");
-    public static final Position D1 = Position.of("d1");
-    public static final Position D2 = Position.of("d2");
-    public static final Position D3 = Position.of("d3");
-    public static final Position D4 = Position.of("d4");
-    public static final Position D5 = Position.of("d5");
-    public static final Position E1 = Position.of("e1");
-    public static final Position E2 = Position.of("e2");
-    public static final Position E3 = Position.of("e3");
-    public static final Position E4 = Position.of("e4");
-    public static final Position E5 = Position.of("e5");
-    public static final Position F3 = Position.of("f3");
-    public static final Position F6 = Position.of("f6");
-    public static final Position G3 = Position.of("g3");
-    public static final Position G7 = Position.of("g7");
-    public static final Position H3 = Position.of("h3");
-    public static final Position H8 = Position.of("h8");
+	public static final Position A1 = Position.of("a1");
+	public static final Position A2 = Position.of("a2");
+	public static final Position A3 = Position.of("a3");
+	public static final Position A4 = Position.of("a4");
+	public static final Position A5 = Position.of("a5");
+	public static final Position B1 = Position.of("b1");
+	public static final Position B2 = Position.of("b2");
+	public static final Position B3 = Position.of("b3");
+	public static final Position B4 = Position.of("b4");
+	public static final Position B5 = Position.of("b5");
+	public static final Position B6 = Position.of("b6");
+	public static final Position C1 = Position.of("c1");
+	public static final Position C2 = Position.of("c2");
+	public static final Position C3 = Position.of("c3");
+	public static final Position C4 = Position.of("c4");
+	public static final Position C5 = Position.of("c5");
+	public static final Position C6 = Position.of("c6");
+	public static final Position C7 = Position.of("c7");
+	public static final Position C8 = Position.of("c8");
+	public static final Position D1 = Position.of("d1");
+	public static final Position D2 = Position.of("d2");
+	public static final Position D3 = Position.of("d3");
+	public static final Position D4 = Position.of("d4");
+	public static final Position D5 = Position.of("d5");
+	public static final Position E1 = Position.of("e1");
+	public static final Position E2 = Position.of("e2");
+	public static final Position E3 = Position.of("e3");
+	public static final Position E4 = Position.of("e4");
+	public static final Position E5 = Position.of("e5");
+	public static final Position F3 = Position.of("f3");
+	public static final Position F6 = Position.of("f6");
+	public static final Position G3 = Position.of("g3");
+	public static final Position G7 = Position.of("g7");
+	public static final Position H3 = Position.of("h3");
+	public static final Position H8 = Position.of("h8");
 
 }
